+++ amazon-kinesis-client/src/main/java/software/amazon/kinesis/coordinator/CoordinatorFactory.java
@@ -22,14 +22,68 @@ import software.amazon.kinesis.leases.ShardInfo;
 import software.amazon.kinesis.processor.Checkpointer;
 
 /**
- *
+ * Used in the process of configuring and providing instances to the {@link Scheduler}
  */
 public interface CoordinatorFactory {
+    /**
+     * Creates the executor service to be used by the Scheduler.
+     *
+     * @return ExecutorService
+     */
     ExecutorService createExecutorService();
 
-    GracefulShutdownCoordinator createGracefulShutdownCoordinator();
+    /**
+     * Creates GracefulShutdownCoordinator to be used by the Scheduler.
+     *
+     * <h3>Method Deprecated</h3>
+     * <p>
+     *     <strong>Note: This method has been deprecated, and will be removed in a future release. Use the configuration in
+     *     {@link CoordinatorConfig#gracefulShutdownCoordinator}. Set the
+     *     {@link CoordinatorConfig#gracefulShutdownCoordinator} to null in order to use this method.</strong>
+     * </p>
+     * <h4>Resolution Order</h3>
+     * <ol>
+     *     <li>{@link CoordinatorConfig#gracefulShutdownCoordinator()}</li>
+     *     <li>{@link CoordinatorFactory#createGracefulShutdownCoordinator()}</li>
+     * </ol>
+     *
+     *
+     * @return a {@link GracefulShutdownCoordinator} that manages the process of shutting down the scheduler.
+     */
+    @Deprecated
+    default GracefulShutdownCoordinator createGracefulShutdownCoordinator() {
+        return new GracefulShutdownCoordinator();
+    }
 
-    WorkerStateChangeListener createWorkerStateChangeListener();
+    /**
+     * Creates a WorkerStateChangeListener to be used by the Scheduler.
+     *
+     * <h3>Method Deprecated</h3>
+     * <p>
+     *     <strong>Note: This method has been deprecated, and will be removed in a future release. Use the configuration in
+     *     {@link CoordinatorConfig#workerStateChangeListener}. Set the
+     *     {@link CoordinatorConfig#workerStateChangeListener} to null in order to use this method.</strong>
+     * </p>
+     *
+     * <h4>Resolution Order</h3>
+     * <ol>
+     *     <li>{@link CoordinatorConfig#workerStateChangeListener()}</li>
+     *     <li>{@link CoordinatorFactory#createWorkerStateChangeListener()}</li>
+     * </ol>
+     *
+     * @return a {@link WorkerStateChangeListener} instance that will be notified for specific {@link Scheduler} steps.
+     */
+    @Deprecated
+    default WorkerStateChangeListener createWorkerStateChangeListener() {
+        return new NoOpWorkerStateChangeListener();
+    }
 
+    /**
+     * Creates a RecordProcessorChedckpointer to be used by the Scheduler.
+     *
+     * @param shardInfo ShardInfo to be used in order to create the ShardRecordProcessorCheckpointer
+     * @param checkpoint Checkpointer to be used in order to create Shardthe RecordProcessorCheckpointer
+     * @return ShardRecordProcessorCheckpointer
+     */
     ShardRecordProcessorCheckpointer createRecordProcessorCheckpointer(ShardInfo shardInfo, Checkpointer checkpoint);
 }