+++ amazon-kinesis-client/src/main/java/software/amazon/kinesis/lifecycle/ShardConsumer.java
@@ -43,6 +43,7 @@ import software.amazon.kinesis.lifecycle.events.ProcessRecordsInput;
 import software.amazon.kinesis.metrics.MetricsCollectingTaskDecorator;
 import software.amazon.kinesis.metrics.MetricsFactory;
 import software.amazon.kinesis.retrieval.RecordsPublisher;
+import software.amazon.kinesis.retrieval.RetryableRetrievalException;
 
 /**
  * Responsible for consuming data records of a (specified) shard.
@@ -241,7 +242,12 @@ public class ShardConsumer {
             return null;
         }
         if (failure != null) {
-            log.warn("{}: Failure occurred in retrieval.  Restarting data requests", shardInfo.shardId(), failure);
+            String logMessage = String.format("%s: Failure occurred in retrieval.  Restarting data requests", shardInfo.shardId());
+            if (failure instanceof RetryableRetrievalException) {
+                log.debug(logMessage, failure.getCause());
+            } else {
+                log.warn(logMessage, failure);
+            }
             startSubscriptions();
             return failure;
         }
@@ -265,6 +271,11 @@ public class ShardConsumer {
                     if (subscriber != null) {
                         subscriber.cancel();
                     }
+                    //
+                    // Set the last request time to now, we specifically don't null it out since we want it to trigger a
+                    // restart if the subscription still doesn't start producing.
+                    //
+                    lastRequestTime = Instant.now();
                     startSubscriptions();
                 }
             }
@@ -294,7 +305,9 @@ public class ShardConsumer {
             if (lastDataArrival != null) {
                 Instant now = Instant.now();
                 Duration timeSince = Duration.between(subscriber.lastDataArrival, now);
-                log.warn("Last time data arrived: {} ({})", lastDataArrival, timeSince);
+                if (timeSince.toMillis() > value) {
+                    log.warn("Last time data arrived: {} ({})", lastDataArrival, timeSince);
+                }
             }
         });
     }