+++ amazon-kinesis-client/src/test/java/software/amazon/kinesis/leases/HierarchichalShardSyncerTest.java
@@ -42,7 +42,7 @@ import java.util.UUID;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;